Add scanner random coincidence scaler to mapfiles.
This signal has been physically present since sometime during the day
2010-10-27, as documented in hclog 202835, and so should be counted
starting from run 6289 or so.

The complaints about the scanner tab crashing the analyzer GUI began
after I added wilbur and charlotte to two of these files but not the
third.  But apparently the scanner tab is still crashing the GUI, so
perhaps someone with a clue can take a look.

git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2153 b9700a98-9937-0410-b777-f318d6eb5bff