Charge Feedback Updates
command line/configuration file flag to set the have-wave plate state. This will switch the PITA slope depending on the plate status. The default setting is Half wave plate OUT to set it to Half wave plate IN use the flag
	--Half-wave-plate-IN

Added two parameters to set the PITA slope for
	1. Half wave plate IN ->  PITASLOPE_IN
	2. Half wave plate OUT -> PITASLOPE_OUT

Added t_0 values for Pockels Cell (PC) +/-. If these two values are set in the parameter file, they will be the initial hardware set point for the PC+/-. Two parameters are,
	1.PC_pos_t0 = 0
	2.PC_neg_t0 = 0
Currently they are set to zero but will be set to correct values later. t_0 correction will not happen as long as they are set to zeros.

PITA correction for PC+/- is only done if the corrected hardware values are within the given limit. Limits are defined in the parameter file.
	PC_low = 38018
	PC_up  = 48016
Otherwise no correction is done at present. But this can be setup to apply the t_0 correction instead of not applying any correction.
	
To-Do
Before running we need to verify the PITA slopes are correct for each half wave plate state.

Rakitha


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2177 b9700a98-9937-0410-b777-f318d6eb5bff