Bug fix for tracking analyzer.  In some event, 0.2%, there was a crash due to my previous memory leak fixes.  This now fixes that crash (and reintroduces a small memory leak):
After 10000 events, number of tracking objects still alive:
  QwEvent: 0
  QwEventHeader: 0
  QwHit: 55
  QwHitPattern: 0
  QwTreeLine: 5
  QwPartialTrack: 0
  QwTrack: 0

I'll see if I can find them, but at least this should be fine now to get
a job to run over an entire tracking data file without either exhausting
the memory or crashing.

I realize that things are running slightly slower now.  This is due to
the copy constructors that are invoked where we used to just store refs.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2184 b9700a98-9937-0410-b777-f318d6eb5bff