Note: this is the branch wdconinc.first_pass_regression.
The regressed variables (and ONLY the regressed variables) are now
written with the same name to a separate tree in the same output
ROOT file.  The trees can be friended.  Someone who has more of a
working knowledge with ROOT can probably figure out a way to get
the trees to 'overlap', using aliases or so.  The original Hel_Tree
is completely untainted by this regression.

Since we seem to have (de facto) standardized on regression of the
asymmetries by the differences, I have removed all other 'example'
regression strategies from the parameter file and executable.  The
numbers themselves in the parameter file are meaningless.

This whole thing hasn't been benchmarked against any of the other
regression codes we have.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/wdconinc.first_pass_regression@2213 b9700a98-9937-0410-b777-f318d6eb5bff