Added support for octo- and hexo-quads to the analyzer.  This was painless.
As with the other helicity parameters, you can set the values in the helicity
map file by specifying (e.g. for quartets, the default)
  PatternPhase=4
  PatternBits=0x69
or by specifying an overriding command line options (e.g. for octoquads)
  --helicity.bitpattern 0x66669999 --helicity.patternphase 32

Note:
  0x6 = binary 0110 = quartet
  0x9 = binary 1001 = quartet
  0x69 = binary 01101001 = octet

The least significant bit is always the earliest helicity window(*).  The input
is flipped internally to start with plus, so 0x69 and 0x96 octets are the same.
The helicity sign is also consistent with what we had before.

(*) We only use symmetric patterns, as far as I know, so this is somewhat moot.

This also works for hexo-quads, but not for deka-quads or dodeka-quads.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2239 b9700a98-9937-0410-b777-f318d6eb5bff