Added a Slow_Tree with EPICS data (very little events).
Big fat disclaimer:
  DONT TRUST EPICS ON EVENT BY EVENT LEVEL.

Now that we have that out of the way, EPICS vars are written to a tree that
can be friended (eh... not quite yet) and used to plot quantities versus EPICS
variables.  Runs start with an EPICS event, so there should not be any period
without valid EPICS data.  This is not intended for constantly varying EPICS
variables, but for EPICS vars that are binary/flags/integer and are constant
for the entire run (e.g. IHWP, XOFF, XSOF, XGOF,...).  This just removes the
need to go and look in the hclog run start/stop entries.

Also:
- removed some compiler warnings from Compton electron detector class,
- some general cleanup of the EPICS class, removed redundant definitions of
  class methods with default arguments, made some properties static.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2253 b9700a98-9937-0410-b777-f318d6eb5bff