Now histograms are in the same level as the running sum in terms of the event cuts. So histograms will not get any events that failed the global event cut. Event ring parameters are adjusted to prevent the events from the current ramp after a beam trip. These values are set as a hard limit on the ring but can be adjusted before a run by using following command line arguments,
  --ring.size arg (=1200)                 QwEventRing: ring/buffer size
  --ring.bt arg (=4)                      QwEventRing: minimum beam trip count
  --ring.hld arg (=52000)                 QwEventRing: ring hold off
  
  
We will eventually try to set up event hold off limit based on the current ramp and the stable beam current. That will be a future update.

The current hold off is set for a current ramp of 0.003uA/events and production current of 150uA. This was the lowest value we observed (run 7729).
Rakitha


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2264 b9700a98-9937-0410-b777-f318d6eb5bff