Added calibration factors for the QPD class to convert ADC counts in to mm to obtain the X and Y position of the laser beam using the four PDs.
The calibration factors are set in the  injector_beamline_geometry.map  which is the "injector" version of the qweak_beamline_geometry.map. It will have the injector beam elements (currently only bpms and qpds) offsets, z locations, gain factor and scaling factors. For the QPDs, the calibration factors for the counts -> mm in X and Y is set in AlphaX and AlphaY. 


The changes in the QwBeamLine class are done in the LoadGeometryDefinitions() function to load geometry for the QPDs.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2301 b9700a98-9937-0410-b777-f318d6eb5bff