In QwQPD, QwLinearDiodeArray and QwBPMCavity classes added "EffectiveCharge" to be written to the data base.
Noticed that the "EffectiveCharge" was being saved in the database as a bpm measurement ie as measurements d and yp. Found out this was because they were being saved in to the data base as bpm measurements inside QwBeamLine class. Changed the measurement type to measurement_type_bcm. This change was done to, BPMs, CombinedBPMs, QPD s and LinearDiodeArray device types. 




git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2309 b9700a98-9937-0410-b777-f318d6eb5bff