Fixed a calculation error in the X position calculation in QPD class. Use fAbsPos in QPD calss to store the beam position in units of mm and fRelPos to store the position in ADC counts. This makes it easier to check the calibration factor used in QPD. Since the QPD calibration factor is applied to ADC counts after some manipulations to get X and Y, we can't used .hw_sum_raw for the same purpose.
Added the QPD running sums to the AccumulateRunningSum() in QwBeamLine class.

Added PrintInfo(), PrintValue(), CalculateRunningAverage(), AccumulateRunningSum() functions to VQwBPM class.  Right now, only the QwQPD class takes full advantage of having the VQwBPM class. Later I need to make sure that the rest of the bpm classes also use this class for the purpose it was created for: a base class for all the bpms to handle absolute psoition and effective charge operations. 

Added the X claibration factor for QPD from the new calculaitons. Y was the same because I made the mistake in the X calculation.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2326 b9700a98-9937-0410-b777-f318d6eb5bff