Now following EPICS variables archive charge asymmetry when feedback program is running.
1. qw:ChargeAsymmetry
2. qw:ChargeAsymmetryError
3. qw:ChargeAsymmetryWidth
4. qw:ChargeFeedback - status of the feedback 1 is ON: 0 is OFF
With the exception
If feedback is killed using 
	kill -9 proc_num
	
status will not be updated to zero. This is updated to zero only with ctrl+x or kill proc_number

Still log files (Feedback_PITA_log.txt and Feedback_IA_log.txt) are active and can be accessed with values from the most recent run.


Added these flags,
  --Half-wave-plate-IN  Set Half wave plate IN. The default is Half wave plate 
                        OUT
  --PITA-Feedback       Run the PITA charge feedback
  --IA-Feedback         Run the PITA charge feedback
  
  Added the following entries to the qwfeedback.conf
  	PITA-Feedback    = true
	IA-Feedback      = false

By default it will run PITA-Feedback and IA-Feedback is disabled.

Rakitha



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2349 b9700a98-9937-0410-b777-f318d6eb5bff