Added scan data (1&2) and clean data decoding functionality to QwEventBuffer class. This will enable us to see these values in tracking runs. The entries in the event_tree are,
1. Coda_CleanData
2. Coda_ScanData1
3. Coda_ScanData2

With this update there will be redundancy in scan data values in the parity runs. This is due to QwHelicity class too decoding scan data values. I will remove this in a later update.

In order to get scan data values in to the tracking CODA data stream, the helicity prescales must not be set to maximum. 


Rakitha


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2355 b9700a98-9937-0410-b777-f318d6eb5bff