Addition of abstract mix-in template class MQwCloneable and additions to the subsystem factory.
Until now neither of the following worked:
  VQwSubsystemParity* lumi1 = new QwLumi("lumi");
  VQwSubsystemParity* lumi2 = lumi1->Copy();
  // fails because Copy returns VQwSubsystem*
and
  VQwSubsystem* lumi1 = new QwLumi("lumi");
  VQwSubsystem* lumi2 = lumi1->Copy();
  // fails because Copy is only in VQwSubsystemParity
It is easy to solve the first one by having Copy return VQwSubsystemParity*.  The second problem can be solved by moving the Copy method into VQwSubsystem.  The second approach seems the cleanest as it provides all subsystems with an easy cloning method.  The Copy() method is used in the subsystem array and has to succeed for every subsystem.

Now, how to move up the Copy() call into VQwSubsystem as painlessly as possible (for subsystem developers)?  Back to our curiously recursive template pattern :-)  We create a virtual mix-in class Cloneable that provides templated cloning functionality.  This additionally allows to keep the subsystem factories inside their class as a static instead of just free floating globals to which the original reference is eternally lost.  The approach with Cloneable requires that we have good copy constructors (as is generally strongly encouraged).

Each subsystem should inherit from MQwCloneable, e.g. class QwLumi: public VQwSubsystemParity, public MQwCloneable<QwLumi>, and the subsystem factory registraction should change to (sorry for the additional line):
  template<> const VQwSubsystemFactory*
  MQwCloneable<QwLumi>::fFactory = new QwSubsystemFactory<QwLumi>("QwLumi");
The Copy() method that each subsystem implements is then not necessary anymore.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2368 b9700a98-9937-0410-b777-f318d6eb5bff