Added additional initialization and error checking to QwEPICSEvent::FillSlowControlsSettings().  Previously, it did not check to see if a given EPICS variable requested via FindIndex() existed in the array before attempting to retrieve it from fEPICSCumulativeData[].  This, coupled with the fact that the current EpicsTable.map on the cdaq cluster is out of date, was causing the seg-faults over the weekend.
Current version of Analysis/prminput/EpicsTable.map on cdaq cluster is r2316 whereas the most up-to-date version dates from r2323.  The current version is missing the declaration of QWtgt_name.  The file must have gotten borked over the weekend; it had been working previously because that field was filled properly in the database up until run 9049.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2407 b9700a98-9937-0410-b777-f318d6eb5bff