Changes to the Compton electron detector map files to calculate asymmetries for the accumulator strip maps.
  qwcompton -r 3811 --detectors compton_electron_detector.map

This required a change to the behavior of userbit helicity mode: before 
this commit the 'scalercounter' field was required to contain the value 
1 (with the intention that this is a scaler channel with the helicity 
trigger as input), now also the value zero is allowed (when no such 
scaler channels is connected).  The error mode that this counter is 
intended to catch is missed helicity windows, when it would count more 
than 1.  This helicity readout mode is not used anymore, so this is 
mostly irrelevant to Qweak but I thought I'd point it out anyway...


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2427 b9700a98-9937-0410-b777-f318d6eb5bff