Added the injector high current calibration pedestals (run 1960 with 140 - 160uA current) taken during January 6th beam studies. To my knowladge, the F2 cup was inserted for this so only bpms starting from 1i02 to 0i05 will have good calibrations. The rest of them wil have low current calibrations from the previous pedestal run (low current). We need to reanalize the runs we are taking now with injector. I will do that.
Note: bcm0l02 will not be usable as it is after F2. I used the EffectiveCharge of bpm 1i02 as a reference bcm to get the pedestals. Checked by reanalyzing run 2001 and the position values seems reasonable now. Earlier they were returning some unreasonable values and it took me a while to realize what may be the cause of that.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2469 b9700a98-9937-0410-b777-f318d6eb5bff