Created a seperate directory in the Macros directory to store the analysis scripts that are needed for transverse data analysis. Following scripts are included:
md_correlations.cc - This script will plot the main detector md pos, md neg and md bar sum asymmetry correlations w.r.t. either position differences or charge asymmetry or energy change. The user has the option of selecting which hs/she would want to use and then they can also select which device they want to use. Ofcourse there maybe simmilar scripts available but I wanted to create a seperate one for transverse measurement so that I can modify it seperatedly as needed.

TransversFit.cc - This script plots md pos, md neg, md bar sum asymmetry vs octant number. It creates 6 plots on the canvas. 3 plots on the left are fitted with a fit of the form a + b* cos(phi+ c) and the plots on the right are fitted with a constant.  For this script the only input required is the run number.

slug_average_fits.cc - This script is used to get the slug averaged md pos, md neg, md bar sum from the qw_fall2010_20101204 data base for the two half wave plate states IN and OUT. There are 3 plots in the canvas for mdpos, mdneg and mdbarsum. Each plot has two graphs and fits for IHWP IN and OUT. The fit I used here is also a + b * cos(phi +c). I am using the cos fit because the up comming transverse measurement will be done using vertical transverse polarization. It is easy to change to sin fit but that modification needs to be done inside the code itself.

All of the scripts needs to be compiled to created there respective executibles and this is done by the  Makefile. 

Well, unfortunately I won't be here to see how these scripts perform. But I hope they will be very usefull to the shift crew for quick analysis. Except for the slug_average_fits which takes a quite a bit of time to get the data from the data base, the other scripts are quick and easy. All the canvases from all the scripts are also saves in to .gif files inside the Transverse directory itself.




git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2501 b9700a98-9937-0410-b777-f318d6eb5bff