Finished implementing consecutive and arbitrary runlet accumuation in histogram mode.
One can now select arbitrary runlets wen typing
in a run number (without resorting to the file
dialog box ...). The runlets are filled in 
automatically, according to the number of runlets
found in the directory. 

Added a few vital plots to the main (front) tab
(more possibly comming ...)

Fixed a number of annoying problems ...





git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2505 b9700a98-9937-0410-b777-f318d6eb5bff