Improved IHWP flip detection in the feedback. The previous IHWP state is saved in a text file
/local/scratch/qweak/Feedback_IHWP.txt


Then this is compared with the current IHWP state to determine the flip has occurred. If IHWP state is flipped, then PC hardware values are updated with nominal values which are saved in the parameter file.

prminput/qweak_fb_prm.in

Currently nominal values are determined by observing the mean hardware value when charge asymmetry is ~0ppm. This has to be updated regularly if the nominal values are shifting. 

I also updated the hall beamline pedestal file for the feedback engine with the latest file in the parity engine,

Parity/prminput/qweak_hallc_pedestal.9695-.map -> Extensions/QwFeedback/prminput/qweak_hallc_pedestal.map



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2545 b9700a98-9937-0410-b777-f318d6eb5bff