Change to the basic qweak_maindet.map file to have tube names in lowercase so it agrees with the root tree variables names. This was simply to make using the setWeights.pl script easier.
setWeights.pl has a description of how to use it at the top of the file. It takes in a textfile containing tube yields in the format seen in the script description as well as at ~/users/smacewan/weights.txt on the cdaq cluster. It uses qweak_maindet.map as a template and create a new mapfile with weights determined using 1/Y_i.

qweak_maindet.10213-.map is the first mapfile generated using this script. I've tested it on the ifarm and it appears to be working fine. More comparisons and weighting tests to be done over the next couple of days.






git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2566 b9700a98-9937-0410-b777-f318d6eb5bff