Added two command line arguments to the update_cache_links perl script.  It now accepts -s <stem> to specify the stem for the link files -r <regex> to specify the regex for the cached data files
Both of those changes were motivated by the Compton: we use Compton_<run>.dat.<segment>,
and the Moller data files are cached in the same directory as the Compton data files, so
we want to avoid conflicts by using a regex.  The shell script can be used to reuse those
arguments repeatedly.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2571 b9700a98-9937-0410-b777-f318d6eb5bff