added generation of tree for regressed DVs, the output tree name is ~ regR10027.000/reg_QwPass1_10027.000.root
It contains: patternNumber, regGlobErrorCode==0 only for good events passing all cuts from the regression code. This cut includes: ErrorFlag==0, bcm1> avr-5uA, and AND on DeviceError for all IVs, DVs, and BCMs read in by the regression code - it is very conservative cut.

The swiss-knife script running 2 regression passes was added. It is enough to execute once per run segment:
 ./doubleRegPass.sh 10027 0

to produce all output files for the run 10027.000 and get them saved in the destination direcotry specified in this script
  



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/balewski.correlator@2578 b9700a98-9937-0410-b777-f318d6eb5bff