Added in some mapfiles with weightings appropriate for some more runs from 2010. I decided that insertion/removal of the octant 7 plug, as well as any high voltages changes, warrented a new set of weights. I also fixed up the setWeights script a bit with some helpful suggestions from Rob (thanks!)
For those following the weighting saga. Note that the trunk version of the analyzer DOES have weights from the "new" scheme wherein we weight solely by yields and not asymmetry widths as well.

The over-attentive collaborator might notice a drastic change in the magnitude of the weights. On this point I would remind him/her that the analyser normalizes by the sum of the weights and that it is the relative size of weights between tubes that matters and not the absolute scale.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2583 b9700a98-9937-0410-b777-f318d6eb5bff