Note to analyzers:   Changes in the behavior of boolean options.  This should lead to more clarity   in the behavior of the analyzer.  You will have to specify true or false in some   cases where you have not had to specify that before.
Note to programmers:
  Please do not use the 'zero_tokens' semantic specifier anymore when you add new
  command line options!  The behavior of zero_tokens is not intuitive to the user.

Basically, the problem we noticed (in the normalization of the lumi, but also in
the magnetic field map disabling), is that you cannot override a boolean option 
that is specified in a configuration file (e.g. if the qwparity.conf file contains
the line QwLumi.normalize = true, then you can specify --QwLumi.normalize=false but
it is still going to normalize).  This is caused by the zero_tokens() semantic 
specifier which just throws away the option value of --QwLumi.normalize and gives 
it the default value.

I removed all occurrences of zero_tokens() on boolean command line options (and on
no other command line options does zero_tokens() make sense at all).  This means that
you WILL HAVE TO EXPLICITLY SPECIFY TRUE OR FALSE, always!

Ok, not always...  In boost::program_options 1.35.0 another semantic specifier was 
added: implicit_value(b).  This allows for dropping the value (and then using the
specified value), but if an explicit value is specified then that value will be used.  
The ifarml machines and RHEL have boost::program_options 1.33.1, so you cannot use
this on those hosts.

Where zero_tokens() and default_value(b) were used before, we now use implicit_value(b)
on systems that support it.  On other systems implicit_value(b) is not used, naturally.
For cases where you used to use zero_tokens, please use default_bool_value(b) with b
the boolean value you want as default (the implicit value will be true, which is the
natural thing for boolean flags).

E.g.:
* enable-runningsum is defined as po::value<bool>()->default_bool_value(true), so if
  not specified this flag is true.  If --enable-runningsum=false is specified, then
  the flag will be false.  If only --enable-runningsum is specified on systems > 1.35.0,
  then the flag will be true (on other systems it will complain and stop).
* enable-burstsum is defined as po::value<bool>()->default_bool_value(false), so if
  not specified this flag is false.  If --enable-burstsum=true is specified, the
  the flag will be true.  If only --enable-burstsum is specified on systems > 1.35.0,
  then the flag will still be true (on other systems it will complain and stop).
* if print-runningsum=true is set in a configuration file, and on the command line you
  specify --print-runningsum=false, then the flag will be set to false.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2619 b9700a98-9937-0410-b777-f318d6eb5bff