Changed mdlumi_ped to save plots and pedestal files into $QWSCRATCH/calib/. We do not need to keep the plots under version control of course, and we may not want to have the pedestal files kept either; putting them in the scratch directory seems to be a better idea than inside the macros directory. I also changed the expected filename for this script. It now looks for files of the format 'qwick_mdlumiped_RUNNUM.root'. These files are generated by calling 'qwick_pedestals' on a cdaq machine. qwick_pedestals is essentially a copy of qwick the qwick analysis but produces files named such that the rootscript can easily grab them.
getpedestals is a perl script that I wrote after John Leacock and I hashed out the details of how we want shift workers to be able to create new pedestal files. It is run by running 'getpedestals RUNNUM' from the Extensions/Macros/Parity/Maindet directory (I might set up a symlink in /home/cdaq/bin for this later) and does the following:

>Creates $QWSCRATCH/calib/pedPlots if it doesn't exist.
>Run qwick_pedestal analysis
>Call mdlumi_ped
>Display reference plots and new plots from qwick analysis
>Asks user if we need to update the pedestal files (for md and lumi subsystems individually)
>If so, moves files to prminput
>close the run range of the last pedestal file
>performs appropriate svn 'add's and 'move's
>emails John and I the plots (regardless of whether or not user decided to update)
>post plots to HCLog (only if user decided to update)

The script still needs to be tested a bit more on the cdaq cluster. I'll put in a wiki page about it when it's ready to be used by shift crews. We also still need to determine when it is neccessary to update pedestal files (we need to make it as clear as possible for the shift crews).


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2634 b9700a98-9937-0410-b777-f318d6eb5bff