We still have the duplicate geometry definition in a couple of places: - the new geometry file qweak_new.geo read by the tracking subsystems, and - the old geometry file qweak.geo read by the Qset object. The reason why we haven't removed this duplication is that we looked at detectors as stand-alone objects, and questions like "is this the first HDC plane?" were impossible to answer like that.
This commit introduces a QwGeometry container class of pointers to QwDetectorInfo
objects.  It has support for a couple of methods to select subsets of detectors,
e.g. In(EQwRegion), In(EQwDetectorPackage), In(EQwDirection), Of(EQwDetectorType).
It also is always sorted by package,region,z-position (operator< on QwDetectorInfo).

This could be pushed out to the tracking subsystems by changing the fDetectorInfo
fields to vectors of QwDetectorInfo objects, instead of vectors of vectors... but
that's not really necessary.

The dependence on Qset and Det will be removed shortly in favor of this new structure.
Wherever the Det.h or Qset.h headers are not necessary they were already removed.

QwTypes: added operator<< for the enums EQwRegion, etc.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2698 b9700a98-9937-0410-b777-f318d6eb5bff