Added routine SetVQWKSaturationLimt(Double_t sat_volts=8.5) to set the saturation limit in volts for VQWK channels. The values is stored in the parameter, fSaturationABSLimit
Currently it is set to 8.5V and can only be set at the compiler time. We could change this if required but for now we have live with that.

Added the error flag,
      kErrorFlag_VQWK_Sat   = 0x01
to identify channels that failed the saturation cut. 

Also I have updated the error summary with the new hardware error flag. 

Tue Mar 22 12:39:12 EDT 2011 Rakitha


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2702 b9700a98-9937-0410-b777-f318d6eb5bff