Store all geometry information for multiple detectors in QwGeometry objects.
The geometry can now be queried from the tracking subsystem array, using GetGeometry().
Each subsystem is responsible for reading the geometry file (as before), figuring out
what lines to use (as before), and creating a new QwDetectorInfo object to be pointed
to from the fDetectorInfo array.  On destruction of the subsystem, all those objects
are nicely deleted.  This all happens in the VQwSubsystemTracking layer.

Some other classes (notably QwTreeEventBuffer for reading Geant4 events) needed to be
made aware of this new structure (because it used to use hard-coded indices).

For all who like to do this, you can again explicitly disable the HDCs when you run
the qwtracking analyzer :-)


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2705 b9700a98-9937-0410-b777-f318d6eb5bff