Note: this is branch pking.templated_beamline.
The BCM and CombinedBCM classes can now use QwScaler_Channel
as the data type.

In the process of making these changes, the functions supporting
the mock data random generator have been moved to a mix-in
class, MQwMockable, from which both QwVQWK_Channel and
QwScaler_Channel inherit.
As yet, I have not tested that the mock data generation works
correctly after this change.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/pking.templated_beamline@2744 b9700a98-9937-0410-b777-f318d6eb5bff