Implementation of a scaler bank class.  This is what has been brought up every so often at the analysis meeting, and would for example combine tracking scalers together with a reference timing channel to produce rates in Hz.  A skeleton was already available, but it didn't do much...
An example channel map for the Compton is provided.  Scalers have to be
specified as 24- or 32-bit.  A normalization channel can be specified,
and is valid until the next normalization channel is specified (or norm
is set to 1).  Normalization channels do not have to precede their use;
that's all figured out after the entire channel map is read in.  They
also don't need to be in the same module (but for DAQ reasons you might
want to avoid cross-module time references.

This commit is one of the building blocks of variable length bursts for
the Compton electron analysis.  The laser power will be retrieved through
publishing of the appropriate scaler channel.

Remaining work:
- allow skipping of channels, right now it (probably?) still needs a
  continuous block of channels...
- publishing of channels for use by other subsystems or the main routine

Other small changes:
- removed unimplemented (and hence unused) functions from a few classes
- error output in VQwSubsystem returns roc and subbank in hex
- debug output in EPICS event removed
- QwScaler_Channel::Normalize(const QwScaler_Channel&)



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2786 b9700a98-9937-0410-b777-f318d6eb5bff