A new qwtracking command line options --QwTracking.print-pattern-db will print the full pattern database used for tracking in HDCs and VDCs (100MB text output!).
For region 2 HDCs it prints the list of nodes, for region 3 VDCs it prints the
actual search tree (the HDC search tree is cyclic, don't know whether to consider
that a bug or whether that's intentional...  it doesn't bother the track finding
itself...).

Sample output up to level 2 of 4 and with 8 wires (for VDCs): see below.  For an
interpretation: rows are wires, columns are bins of VDC timing.  Each ascii image
corresponds to an acceptable hit pattern at that level of resolution.  E.g. the
second pattern does not have any son patterns because at the next level of bin
division there is no possible hit pattern consistent with a straight track that
reduces to this pattern when the resolution is worsened by a factor two.  The
fourth pattern does have a son pattern consistent with a straight track.

Command line:
  qwtracking -r 8658 -e 1 --QwTracking.disable-momentum yes --QwTracking.print-pattern-db

Trees:
tree 0: 0x88cd644: (-1,*) 0,0,0,0,0,0,0,0, width = 1
 |
 |
 |
 |
 |
 |
 |
 |
 son 0: 0x88f9b80: (0,*) 0,1,1,1,1,1,1,1, width = 2
  tree: 0x88db2c4: (0,*) 0,1,1,1,1,1,1,1, width = 2
   |.
   .|
   .|
   .|
   .|
   .|
   .|
   .|
  next: 0x88f9b60: (0,*) 0,0,1,1,1,1,1,1, width = 2
  tree: 0x88db2a0: (0,*) 0,0,1,1,1,1,1,1, width = 2
   |.
   |.
   .|
   .|
   .|
   .|
   .|
   .|
  next: 0x88f9a80: (0,*) 0,0,0,1,1,1,1,1, width = 2
  tree: 0x88db1a4: (0,*) 0,0,0,1,1,1,1,1, width = 2
   |.
   |.
   |.
   .|
   .|
   .|
   .|
   .|
   son 0: 0x88f9b40: (1,*) 0,1,1,2,2,3,3,3, width = 4
    tree: 0x88db27c: (1,*) 0,1,1,2,2,3,3,3, width = 4
     |...
     .|..
     .|..
     ..|.
     ..|.
     ...|
     ...|
     ...|



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2815 b9700a98-9937-0410-b777-f318d6eb5bff