This commit sets the rotation angles for 3P02A and 3P02B which have different rotation angles than than the standard of 45 degrees. Based on communications with Jay Benesch, I am setting the following values for the angles which we need to rotate these two bpms to get them back to the accelerator coordinates.
IPM3P02B rotation = 0  
IPM3P02A rotation = 29

Jay said ops uses the 45 degree anticlockwise rotated position of the bpm as THET = 0 and they had to set 
 caget IPM3P02B.THET = 16
 caget IPM3P02A.THET = 45 
to fix the installation error on these bpms. This means 3P02B is installed with a 45 degree rotation in the clockwise direction and so we don not need to rotate it back to go to the accelerator coordinates since it is in accelerator coordinates. In the same way, if we want to get bpm3p02A to accelerator coordinates we have to rotate by 45-16 = 29 degrees.


 


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2832 b9700a98-9937-0410-b777-f318d6eb5bff