Changed the default value of the normalization flag for both the main detectors and the lumis to be true, instead of false.
Modified the QwOptions to exit if an unrecognized option
is specified in a configuration file.

Removed the normalization and autoflush options from the
qwparity.conf file; this file is now empty.  The values
of these options had matched the default values.

Corrected a option name in the mdlumiped.conf file.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2847 b9700a98-9937-0410-b777-f318d6eb5bff