Scaler counts are normalized to 4MHz clock scaler to obtain actual rates. So far this nomalization is hard coded (by means of 4MHz scaler name and 4MHz rate.) for beamline scalers. 	rate = scaler_counts*4e6/scaler_4mhz_counts
Added capability to calibrate scalers using a calibration factor and a pedestal. Currently hall A scaler is calibrated to 
provide current and charge asymmetry. This is implemented in the Hall C beamline class. calibration and pedestal are set in 
the Hall C beamline pedestal file.

Added event cut functionality to scalers. Now local or global event cut can be set on scalers. This is implemented to 
the hall A scaler bcm. event cut is set in the Hall C beamline event cut file.

I will update the pedestal and event cut for hall A scaler for the main analyzer in a separate commit.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2850 b9700a98-9937-0410-b777-f318d6eb5bff