Note:  This is in branch pking.templated_beamline.
All of the following is to support swithcing to templated
BPM objects.

Modified the combined BPM and similar classes to use
pointers to VQwBPM as the elements rather than specifying
pointers to QwBPMStripline's.
Made the elements in the combined BPM and similar classes
const pointers.
Changed the array indexing to use a standard BPM axis enumeration
rather than hardcoded integers.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/pking.templated_beamline@2879 b9700a98-9937-0410-b777-f318d6eb5bff