NOTE:  This is on branch pking.templated_beamline.
Introduced a new virtual base class, VQwHardwareChannel,
from which both QwVQWK_Channel and QwScaler_Channel
inherit.
Operations that make sense for a single hardware
channel but don't make sense for a composite data
element will be moved from VQwDataElement into
VQwHardwareChannel.

This is an intermediate commit; the code compiles and
runs, and produces the same ROOT files as the trunk@2880,
but not all of the cleanup has happened yet.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/pking.templated_beamline@2884 b9700a98-9937-0410-b777-f318d6eb5bff