This is a reorganization for the raster VQWK data processing. These changes allow the raster "asymmetries" to be handled as differences.  This should also make it easier to further extend the raster data processing in the engine, if we determine that it is needed.
Changes:
========
The changes to QwBCM which were made in revision 2906 were
rolled back to the 2905 version.

A new subsystem, QwIntegratedRaster, and data element,
QwIntegratedRasterChannel, were introduced.  These are
adapted from the QwLumi and QwBCM classes respectively.

The QwIntegratedRaster and QwBeamLine subsystems use the
same set of parameter files.  This is a choice made to make
the database construction scripts easier; we may decide
to split the parameter files later.
The QwIntegratedRaster subsystem will only process data
elements with the "raster" detector type.
The QwBeamLine subsystem will ignore data elements with
the "raster" detector type, although at this point an
error message is produced.

The detector type for the three raster channels in all of
the beamline channel maps was changed from "bcm" to "raster".

The QwIntegratedRaster header was added to QwParity, and
a initalization block was added to the detectors.map.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2931 b9700a98-9937-0410-b777-f318d6eb5bff