Note:  This is in branch release_1.09.
Modified the qwbatchsub.pl to use long options in
addition to the single charater options.  Added the
long option names in the "usage".

Only one single character option has gone away, but
that option is the "-r <run option>" flag for specifying
the run range; it is replaced by "--runs <run option>".

Added new options to define the post-processing scripts,
and started to add an option to define the rootfiles output
directory path.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/release_1.09@2945 b9700a98-9937-0410-b777-f318d6eb5bff