Finalized the target bpm. The target bpm will be made out of the combination of bpms 7A,7B,7C ,9 and 9B. Due to some odd behaviour in 8 it will not be used for the target bpm calculation for this run. All the weights for the least square fit are set to 1 for all the bpms. Updated all the qweak_hallc_beamline.* map files with this new target bpm.
The bpms 7A,7B and 9  are forced to behave in a simmilar manner as the orbit lock bpms 7c and 9b by using 7c and 9b to project to each one of them and getting the slopes and offsets. The offests are then used instead of the x and y offsets from survey and the 1/slopes are used as relative x/y gains. 
e.g.
absX = gainX (relX+offset)

The offsets of 7c and 9b remains unchanged.
The relative X and Y gains are set in the qweak_beamline_geometry.map file using keywords "xgain" and "ygain". The changes done to the qweak_beamline_geometry.map file are to use this changes.
Created a new function in QwBeamLine class to read these keywords and apply them to the individual bpms. The default values are 1. All the modifications done to the VQwBPM are related to setting the relative gains for the bpms and all the changes done to the QwBPMStripline class are to modify the ProcessEbvent() funciton to use the relative gains. 



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@2952 b9700a98-9937-0410-b777-f318d6eb5bff